Removing --force flag from cld env delete
The force flag is supposed to trigger delete of an environment
even though applications are running on it. The initial implementation
of this flag deletes the ECS cluster. But some of the other
resources, such as security group (env resource) and
ecs service (app resource), need additional work to remove them
while the application is running. So commenting this flag till we
add force termination/removal for all the relevant resources of
the environment and the app running on it.