Using name in all user facing operations
Earlier user facing operations were mix of names and ids.
This was affecting usability since one had to know the id
even when an artifact (environment, app) had been created
with a name.

Now ids are only used internally. User facing operations will
only use names.

We are also not going to display the id in an artifact's response